Fixing a space typo in json schema (#11)
To send a [data message](https://firebase.google.com/docs/cloud-messaging/concept-options#notifications_and_data_messages) from Firebase to Android, `notification` key should be left empty. The typo prevents it, as it sends a request which looks like this:

```json
{
  "registration_ids": ["token1", "token2"],
  "content_available": true,
  "mutable_content": true,
  "notification": {
    "android_channel_id": ""
  },
  "data": {
    "hello": "world"
  }
}
```

`ChannelId` is renamed to `ChannelID`, because of `golint` warning: 
> struct field ChannelId should be ChannelID (golint)

A **gorush** project probably needs a vendor update after that as well.